+++ x/rewards/ante/min_cons_fee.go
@@ -4,8 +4,6 @@ import (
 	"github.com/cosmos/cosmos-sdk/codec"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkErrors "github.com/cosmos/cosmos-sdk/types/errors"
-
-	"github.com/archway-network/archway/pkg"
 )
 
 // MinFeeDecorator rejects transaction if its fees are less than minimum fees defined by the x/rewards module.
@@ -38,20 +36,13 @@ func (mfd MinFeeDecorator) AnteHandle(ctx sdk.Context, tx sdk.Tx, simulate bool,
 	}
 
 	var expectedFees sdk.Coins // All the fees which need to be paid for the given tx. includes min consensus fee + every contract flat fee
-	gasUnitPrice, found := mfd.rewardsKeeper.GetMinConsensusFee(ctx)
-	if found {
-		// Estimate the minimum fee expected
-		// We use RoundInt here since minimum fee must be GTE calculated amount
-		txGasLimit := pkg.NewDecFromUint64(feeTx.GetGas())
-		if txGasLimit.IsZero() {
-			return ctx, sdkErrors.Wrap(sdkErrors.ErrInvalidRequest, "tx gas limit is not set")
-		}
-		minFeeExpected := sdk.Coin{
-			Denom:  gasUnitPrice.Denom,
-			Amount: gasUnitPrice.Amount.Mul(txGasLimit).RoundInt(),
-		}
-		expectedFees = expectedFees.Add(minFeeExpected)
-	}
+	computationalGasPrice := mfd.rewardsKeeper.ComputationalPriceOfGas(ctx)
+	expectedFees = expectedFees.Add(
+		sdk.NewCoin(
+			computationalGasPrice.Denom,
+			computationalGasPrice.Amount.MulInt(sdk.NewIntFromUint64(feeTx.GetGas())).TruncateInt(),
+		),
+	)
 
 	// Get flatfees for any contracts being called in the tx.msgs
 	for _, m := range tx.GetMsgs() {