+++ x/rewards/module.go
@@ -89,7 +89,6 @@ func (a AppModuleBasic) GetQueryCmd() *cobra.Command {
 type AppModule struct {
 	AppModuleBasic
 
-	cdc    codec.Codec
 	keeper keeper.Keeper
 }
 
@@ -172,7 +171,6 @@ func (a AppModule) RandomizedParams(r *rand.Rand) []simTypes.ParamChange {
 
 // RegisterStoreDecoder registers a decoder for the module's types.
 func (a AppModule) RegisterStoreDecoder(_ sdk.StoreDecoderRegistry) {
-
 }
 
 // WeightedOperations returns all the module operations with their respective weights.